Revert "最终分数的换算"
This reverts commit c0b4846c76564793be5b4927ce1dcd016b50b13c.

+++ src/main/java/gov/jiusan/star/score/Score.java
@@ -79,7 +79,7 @@ public class Score implements Serializable {
      * 最终得分
      */
     @Column(name = "final_score")
-    private Double finalScore;
+    private Integer finalScore;
 
     /**
      * 创建时间
@@ -162,11 +162,11 @@ public class Score implements Serializable {
         this.aATotalScore = aATotalScore;
     }
 
-    public Double getFinalScore() {
+    public Integer getFinalScore() {
         return finalScore;
     }
 
-    public void setFinalScore(Double finalScore) {
+    public void setFinalScore(Integer finalScore) {
         this.finalScore = finalScore;
     }
